pyproject.toml
[project]
name = "public-data"
version = "0.1.0"
description = "Public data for Bradford 2025"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mkdocs>=1.6.1",
    "mkdocs-include-markdown-plugin>=7.2.0",
    "mkdocs-jupyter>=0.25.1",
    "mkdocs-macros-plugin>=1.5.0",
    "mkdocs-material>=9.7.0",
    "mkdocs-mermaid2-plugin>=1.2.3",
    "mkdocstrings[python]>=1.0.0",
    "ruff>=0.14.10",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
serve_docs="mkdocs serve --livereload"
build_docs="mkdocs build"
